common.models -> common.schema, join all schema-related models back together
