--- application/blueprints/authentication/forms.py
@@ -1,59 +1,0 @@
-from flask_wtf import FlaskForm
-from wtforms import StringField, PasswordField, SubmitField
-from wtforms.validators import (
-    DataRequired,
-    Email,
-    EqualTo,
-    Length,
-    Optional
-)
-
-
-class RegisterForm(FlaskForm):
-    username = StringField(
-        'Username',
-        validators=[
-            Length(min=5),
-            DataRequired()
-        ]
-    )
-
-    name = StringField(
-        'Name',
-        validators=[Optional()]
-    )
-    email = StringField(
-        'Email',
-        validators=[
-            Length(min=6),
-            Email(message='Enter a valid email.'),
-            DataRequired()
-        ]
-    )
-    password = PasswordField(
-        'Password',
-        validators=[
-            DataRequired(),
-            Length(min=6, message='Select a stronger password.')
-        ]
-    )
-    confirm = PasswordField(
-        'Confirm Your Password',
-        validators=[
-            DataRequired(),
-            EqualTo('password', message='Passwords must match.')
-        ]
-    )
-    submit = SubmitField('Register')
-
-
-class LoginForm(FlaskForm):
-    email = StringField(
-        'Email',
-        validators=[
-            DataRequired(),
-            Email(message='Enter a valid email.')
-        ]
-    )
-    password = PasswordField('Password', validators=[DataRequired()])
-    submit = SubmitField('Log In')
