--- application/blueprints/authentication/models.py
@@ -1,60 +1,0 @@
-from flask_login import UserMixin
-from werkzeug.security import generate_password_hash, check_password_hash
-
-class User(UserMixin, db.Model):
-    __tablename__ = 'flasklogin_users'
-
-    id = db.Column(
-        db.Integer,
-        primary_key=True
-    )
-    username = db.Column(
-        db.String(100),
-        nullable=False,
-        unique=True
-    )
-    name = db.Column(
-        db.String(100),
-        nullable=False,
-        unique=False
-    )
-    email = db.Column(
-        db.String(320),
-        unique=True,
-        nullable=False
-    )
-    password = db.Column(
-        db.String(200),
-        primary_key=False,
-        unique=False,
-        nullable=False
-	)
-    created_on = db.Column(
-        db.DateTime,
-        index=False,
-        unique=False,
-        nullable=True
-    )
-    last_login = db.Column(
-        db.DateTime,
-        index=False,
-        unique=False,
-        nullable=True
-    )
-
-    def set_password(self, password):
-        """Create hashed password."""
-        self.password = generate_password_hash(
-            password,
-            method='sha256'
-        )
-
-    def check_password(self, password):
-        """Check hashed password."""
-        return check_password_hash(self.password, password)
-
-    def get_id(self):
-        return str(self.id)
-
-    def __repr__(self):
-        return '<User {}>'.format(self.username)
