--- application/blueprints/authentication/templates/login.html
@@ -1,54 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-<div class="form-wrapper">
-
-    {% for message in get_flashed_messages() %}
-    <div class="alert alert-warning">
-        <button type="button" class="close" data-dismiss="alert">
-            x
-        </button>
-        {{ message }}
-    </div>
-    {% endfor %}
-
-    <h1>Log In</h1>
-
-    <form method="POST" action="/auth/login">
-        {{ form.csrf_token }}
-
-        <fieldset class="email">
-            {{ form.email.label }}
-            {{ form.email(placeholder='youremail@example.com') }}
-            {% if form.email.errors %}
-            <ul class="errors">
-                {% for error in form.email.errors %}
-                <li>{{ error }}</li>{% endfor %}
-            </ul>
-            {% endif %}
-        </fieldset>
-
-        <fieldset class="password">
-            {{ form.password.label }}
-            {{ form.password }}
-            {% if form.email.errors %}
-            <ul class="errors">
-                {% for error in form.password.errors %}
-                <li>{{ error }}</li>{% endfor %}
-            </ul>
-            {% endif %}
-        </fieldset>
-
-        <div class="submit-button">
-            {{ form.submit }}
-        </div>
-
-        <div class="login-register">
-            <span>Don't have an account?</span>
-            <a href="{{ url_for('authentication.register') }}">Sign up.</a>
-        </div>
-
-    </form>
-</div>
-
-{% endblock %}
