--- application/blueprints/authentication/templates/register.html
@@ -1,95 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-<div class="form-wrapper">
-
-    <div class="logo">
-        <img src="{{ url_for('static', filename='dist/img/logo.png') }}" alt="logo" />
-    </div>
-
-    {% for message in get_flashed_messages() %}
-    <div class="alert alert-warning">
-        <button type="button" class="close" data-dismiss="alert">
-            x
-        </button>
-        {{ message }}
-    </div>
-    {% endfor %}
-
-    <h1>Sign Up</h1>
-
-    <form method="POST" action="/auth/register">
-        {{ form.csrf_token }}
-
-        <fieldset class="username">
-            {{ form.username.label }}
-            {{ form.username }}
-            {% if form.username.errors %}
-            <ul class="errors">
-                {% for error in form.username.errors %}
-                <li>{{ error }}</li>
-                {% endfor %}
-            </ul>
-            {% endif %}
-        </fieldset>
-
-        <fieldset class="name">
-            {{ form.name.label }}
-            {{ form.name(placeholder='John Smith') }}
-            {% if form.name.errors %}
-            <ul class="errors">
-                {% for error in form.email.errors %}
-                <li>{{ error }}</li>
-                {% endfor %}
-            </ul>
-            {% endif %}
-        </fieldset>
-
-        <fieldset class="email">
-            {{ form.email.label }}
-            {{ form.email(placeholder='youremail@example.com') }}
-            {% if form.email.errors %}
-            <ul class="errors">
-                {% for error in form.email.errors %}
-                <li>{{ error }}</li>{% endfor %}
-            </ul>
-            {% endif %}
-        </fieldset>
-
-        <fieldset class="password">
-            {{ form.password.label }}
-            {{ form.password }}
-            {% if form.password.errors %}
-            <ul class="errors">
-                {% for error in form.password.errors %}
-                <li>{{ error }}</li>{% endfor %}
-            </ul>
-            {% endif %}
-        </fieldset>
-
-        <fieldset class="confirm">
-            {{ form.confirm.label }}
-            {{ form.confirm }}
-            {% if form.confirm.errors %}
-            <ul class="errors">
-                {% for error in form.confirm.errors %}
-                <li>{{ error }}</li>{% endfor %}
-            </ul>
-            {% endif %}
-        </fieldset>
-
-        <div class="submit-button">
-            {{ form.submit }}
-        </div>
-
-    </form>
-
-    <div class="login-register">
-        <div>
-            <span>Already have an account?</span>
-            <a href="{{ url_for('authentication.login') }}">Log in.</a>
-        </div>
-    </div>
-</div>
-
-{% endblock %}
